"browser" in package.json to prevent "crypto" shim
