+++ lib/sha256.js
@@ -4,54 +4,13 @@
 var _imports = {}
 
 if (typeof module !== 'undefined' && module.exports) { //CommonJS
-  if (typeof process !== 'undefined' && process.pid) {
-    // Node.js
-	module.exports = sha256_node;
-  } else {
-    _imports.bytesToHex = require('convert-hex').bytesToHex
-    _imports.convertString = require('convert-string')
-    module.exports = sha256
-  }
+  _imports.bytesToHex = require('convert-hex').bytesToHex
+  _imports.convertString = require('convert-string')
+  module.exports = sha256
 } else {
   _imports.bytesToHex = globals.convertHex.bytesToHex
   _imports.convertString = globals.convertString
   globals.sha256 = sha256
-}
-
-
-// Node.js has its own Crypto function that can handle this natively
-function sha256_node(message, options) {
-	var crypto = require('crypto');
-	var c = crypto.createHash('sha256');
-	
-	if (Buffer.isBuffer(message)) {
-		c.update(message);
-	} else if (Array.isArray(message)) {
-		// Array of byte values
-		c.update(new Buffer(message));
-	} else {
-		// Otherwise, treat as a binary string
-		c.update(new Buffer(message, 'binary'));
-	}
-	var buf = c.digest();
-	
-	if (options && options.asBytes) {
-		// Array of bytes as decimal integers
-		var a = [];
-		for(var i = 0; i < buf.length; i++) {
-			a.push(buf[i]);
-		}
-		return a;
-	} else if (options && options.asString) {
-		// Binary string
-		return buf.toString('binary');
-	} else {
-		// String of hex characters
-		return buf.toString('hex');
-	}
-}
-sha256_node.x2 = function(message, options) {
-	return sha256_node(sha256_node(message, { asBytes:true }), options)
 }
 
 /*
